Split BasicGraphicsScene group handling
